replaced 'add contact' bubble handler with on*
