+++ src/gui/com/index.js
@@ -91,8 +91,8 @@ var sidehelp =
 exports.sidehelp = function(state, opts) {
   return h('ul.list-unstyled',
     h('li', h('button.btn.btn-primary', { onclick: state.showUserId }, 'Get your contact id')),
-    h('li', h('button.btn.btn-primary.click-add-contact', 'Add a contact')),
-    h('li', h('button.btn.btn-primary.click-add-contact', 'Use an invite')),
+    h('li', h('button.btn.btn-primary', { onclick: state.followPrompt }, 'Add a contact')),
+    h('li', h('button.btn.btn-primary', { onclick: state.followPrompt }, 'Use an invite')),
     (!opts || !opts.noMore) ? h('li', h('span', {style:'display: inline-block; padding: 6px 12px'}, a('#/help', 'More help'))) : ''
   )
 }
